refactor: extract diagnostic checks from diagnostic_dialog.py (641 -> 220 lines)
Moved diagnostic registry, pattern matching, and startup diagnostics
into diagnostic_checks.py. Dialog classes stay in diagnostic_dialog.py.
preferences_dialog.py left as-is (too tightly coupled for clean split).
